+++ src/PSRule.Rules.Azure/Data/Template/IValidationContext.cs
@@ -19,6 +19,12 @@ internal interface IValidationContext
     ResourceProviderType[] GetResourceType(string providerNamespace, string resourceType);
 
     /// <summary>
+    /// Lookup an Azure location by name.
+    /// </summary>
+    /// <param name="location">A normalized name of an Azure location, e.g. "eastus".</param>
+    AzureLocationEntry GetAzureLocation(string location);
+
+    /// <summary>
     /// Determines if the value is a previously identified secure value.
     /// </summary>
     bool IsSecureValue(object value);